Refactor search functionality to custom use hook for modularity
